Extract matrix/transform helpers from main.ts into utils/matrix.ts
Pulls createRotationMatrix, createTranslationMatrix, createQuaternionMatrix,
createAngleAxisMatrix, parseSpaceSeparatedValues, and parseMatrixInput out of
PointCloudVisualizer since they had no dependency on `this` (scene/DOM state).
First step of the main.ts de-bloating effort described in CLAUDE.md.

Also adds eslint.config.js (flat config) mirroring the existing
.eslintrc.json rules: eslint was bumped to v10 (flat-config only) in the
lockfile without migrating the config, which broke the pre-commit hook for
everyone. Same rules, no CI existed to catch it.
